flows/microstrategy_da/requirements.txt
prefect==1.*
boto3==1.24.*
mstrio-py==11.3.*
pandas==1.4.*
